Subindo melhorias

+++ workspace/trabalhoIndividual/montadora/src/main/java/br/com/example/montadora/security/services/CarroServices.java
@@ -7,28 +7,34 @@ import org.springframework.stereotype.Service;
 
 import br.com.example.montadora.security.dto.CarroResponseDTO;
 import br.com.example.montadora.security.entities.Carro;
+import br.com.example.montadora.security.entities.Concessionaria;
 import br.com.example.montadora.security.repositories.CarroRepository;
+import br.com.example.montadora.security.repositories.ConcessionariaRepository;
 
 @Service
 public class CarroServices {
 
 	@Autowired
 	private CarroRepository carroRepository;
+	
+	@Autowired
+	private ConcessionariaRepository concessionariaRepository;
 
 	public List<Carro> listarCarros() {
 		return carroRepository.findAll();
 	}
 
-	public CarroResponseDTO adicionarCarro(CarroResponseDTO carroResponseDTO) {
-		CarroResponseDTO carroDTO = new CarroResponseDTO();
-		carroDTO.setMarca(carroResponseDTO.getMarca());
-		carroDTO.setModelo(carroResponseDTO.getModelo());
-		carroDTO.setAno(carroResponseDTO.getAno());
+	public void cadastrarCarro(CarroResponseDTO carroResponseDTO) {
+		Carro newCarro = new Carro();
+		newCarro.setMarca(carroResponseDTO.getMarca());
+		newCarro.setModelo(carroResponseDTO.getModelo());
+		newCarro.setAno(carroResponseDTO.getAno());
+
+		Concessionaria concessionaria = concessionariaRepository.buscarConce(carroResponseDTO.getNomeConcessionaria());
+		newCarro.setFkConcessionaria(concessionaria);
 		
-		Carro carro = carroResponseDTO.toCarro();
-		carroRepository.save(carro);
+		carroRepository.save(newCarro);
 		
-		return carroDTO;
 	}
 
 	public CarroResponseDTO buscarPorId(Integer id) {
@@ -40,11 +46,26 @@ public class CarroServices {
 		return carroResponseDTO;
 	}
 
-	public void deletarPorId(Integer id) {
-		carroRepository.deleteById(id);
+	public boolean deletarCarro(Integer id) {
+		if(carroRepository.existsById(id)) {
+			carroRepository.deleteById(id);
+			return true;
+		} else {
+			return false;
+		}
 	}
 
-	public Carro atualizar(Carro carro) {
-		return carroRepository.save(carro);
+	public Carro atualizarCarro(Integer id, CarroResponseDTO carroResponseDTO) {
+		Carro carroExistente = carroRepository.findById(id).orElse(null);
+		
+		if(carroExistente != null) {
+			carroExistente.setMarca(carroResponseDTO.getMarca());
+			carroExistente.setModelo(carroResponseDTO.getModelo());
+			carroExistente.setAno(carroResponseDTO.getAno());
+			
+			return carroRepository.save(carroExistente);
+		} else {
+			return null;
+		}
 	}
 }
